fix: 数据源查看隐藏较深优化 数据预览页面前端自适应 SQL 数据集旁边提示增加描述 #3812
